requirements/dev.txt
-r requirements.txt
-r docs.txt
pre-commit
pytest
pytest-asyncio
pillow
datamodel-code-generator[http]
